// react-app/src/Shared/Types.ts
/*
 * @Author: Kanata You 
 * @Date: 2020-11-10 14:46:53 
 * @Last Modified by: Kanata You
 * @Last Modified time: 2020-11-10 15:46:35
 */

export interface Config {
    storagePath: string;
};

export type KeywordType = "keyword" | "type" | "control" | "declare" | "modifier"
    | "operator" | "constance";